csi-wrapper: attach disks in CAA resource group
The documented way for setting up CAA on AKS involves creating the CAA
resources (Pod VM etc.) in a resource group separate from the AKS rg.
This clashes with the azuredisk-csi-driver, which relies on the cloud
config file supplied by the AKS cluster to find the resource group where
disks should be created and attached to VMs.
Using a Kubenerets secret mirroring this cloud config file, we can swap
out the resource group without impacting the rest of the cluster.

Signed-off-by: Daniel Weiße <dw@edgeless.systems>